+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.0.4
 - Add get_trades_history method
-- Unittest were migrated to pytest
+- Auth tests migrated from unittest to pytest
 
 ### 0.0.3
 - Bug fixing